refactoring

+++ inc/unit.h
@@ -2,6 +2,7 @@
 #define __UNIT_H__
 
 #include <fstream>
+#include <stdint.h>
 #include "common.h"
 
 namespace unit_ns
@@ -13,38 +14,38 @@ using std::ostream;
 
 class unit
 {
-	long num{0};
+	int32_t num {0};
 
 public:
 
 	unit();
-	unit(const long &);
+	unit(const int32_t&);
 	~unit() {};
-	unit operator+(const unit &) const;
-	unit operator-(const unit &) const;
-	unit operator*(const unit &) const;
-	unit operator/(const unit &) const;
+	unit operator+(const unit&) const;
+	unit operator-(const unit&) const;
+	unit operator*(const unit&) const;
+	unit operator/(const unit&) const;
 	unit operator-() const;
-	unit &operator+=(const unit &);
-	unit &operator-=(const unit &);
-	unit &operator*=(const unit &);
-	unit &operator/=(const unit &);
+	unit& operator+=(const unit&);
+	unit& operator-=(const unit&);
+	unit& operator*=(const unit&);
+	unit& operator/=(const unit&);
 	operator int();
 	operator long();
-	long convert(const char *);
-	bool read(ifstream &);
+	int32_t convert(const char*);
+	bool read(ifstream&);
 	void print() const;
 
-	friend int operator>(const unit &, const unit &);
-	friend int operator>=(const unit &, const unit &);
-	friend int operator<(const unit &, const unit &);
-	friend int mod(const long &);
-	friend unit abs(const unit &);
-	friend unit sin(const unit &);
-	friend unit cos(const unit &);
-	friend unit sqrt(const unit &);
-	friend ostream &operator<<(ostream &, const unit &);
-	friend istream &operator>>(istream &, unit &);
+	friend int operator>(const unit&, const unit&);
+	friend int operator>=(const unit&, const unit&);
+	friend int operator<(const unit&, const unit&);
+	friend int mod(const int32_t&);
+	friend unit abs(const unit&);
+	friend unit sin(const unit&);
+	friend unit cos(const unit&);
+	friend unit sqrt(const unit&);
+	friend ostream& operator<<(ostream&, const unit&);
+	friend istream& operator>>(istream&, unit&);
 };
 
 const unit UNIT(1024);
